requirements.txt
annotated-types==0.7.0
anyio==4.5.2
blinker==1.8.2
certifi==2025.8.3
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
dlib==19.24.2
exceptiongroup==1.3.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
gotrue==2.9.2
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
numpy==1.24.4
packaging==25.0
pillow==10.4.0
postgrest==0.16.11
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.6
six==1.17.0
sniffio==1.3.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
typing_extensions==4.13.2
websockets==12.0
Werkzeug==3.0.6
zipp==3.20.2